chore: fixes after edc vc module refactor
