+++ core/identity-hub-credentials/src/test/java/org/eclipse/edc/identityhub/core/VerifiablePresentationServiceImplTest.java
@@ -17,12 +17,12 @@ package org.eclipse.edc.identityhub.core;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.json.JsonObject;
-import org.eclipse.edc.iam.identitytrust.spi.model.CredentialFormat;
-import org.eclipse.edc.iam.identitytrust.spi.model.CredentialSubject;
-import org.eclipse.edc.iam.identitytrust.spi.model.Issuer;
-import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredential;
-import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredentialContainer;
-import org.eclipse.edc.iam.identitytrust.spi.model.presentationdefinition.PresentationDefinition;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialFormat;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialSubject;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.Issuer;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.VerifiableCredential;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.VerifiableCredentialContainer;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.presentationdefinition.PresentationDefinition;
 import org.eclipse.edc.identityhub.spi.generator.PresentationCreatorRegistry;
 import org.eclipse.edc.jsonld.util.JacksonJsonLd;
 import org.eclipse.edc.jwt.spi.JwtRegisteredClaimNames;
@@ -33,8 +33,8 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 
-import static org.eclipse.edc.iam.identitytrust.spi.model.CredentialFormat.JSON_LD;
-import static org.eclipse.edc.iam.identitytrust.spi.model.CredentialFormat.JWT;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialFormat.JSON_LD;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialFormat.JWT;
 import static org.eclipse.edc.identityhub.core.creators.TestData.EMPTY_LDP_VP;
 import static org.eclipse.edc.identityhub.core.creators.TestData.JWT_VP;
 import static org.eclipse.edc.identityhub.core.creators.TestData.LDP_VP_WITH_PROOF;
