+++ spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/generator/PresentationGenerator.java
@@ -15,7 +15,7 @@
 package org.eclipse.edc.identityhub.spi.generator;
 
 
-import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredentialContainer;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.VerifiableCredentialContainer;
 
 import java.util.List;
 import java.util.Map;
