+++ spi/identity-hub-spi/src/main/java/org/eclipse/edc/identityhub/spi/generator/VerifiablePresentationService.java
@@ -14,9 +14,10 @@
 
 package org.eclipse.edc.identityhub.spi.generator;
 
-import org.eclipse.edc.iam.identitytrust.spi.model.VerifiableCredentialContainer;
-import org.eclipse.edc.iam.identitytrust.spi.model.credentialservice.PresentationResponseMessage;
-import org.eclipse.edc.iam.identitytrust.spi.model.presentationdefinition.PresentationDefinition;
+
+import org.eclipse.edc.iam.identitytrust.spi.model.PresentationResponseMessage;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.VerifiableCredentialContainer;
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.presentationdefinition.PresentationDefinition;
 import org.eclipse.edc.spi.result.Result;
 import org.jetbrains.annotations.Nullable;
